backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
python-dotenv
jinja2
aiofiles
groq
scikit-learn
pandas
numpy
